Allow #identifier to return identifiers other than wikidata
Currently the #identifier method can only be used to return identifiers
with the scheme 'wikidata'. I'd like to use it to return other id’s.

This commit changes the #identifier method to enable lookups of
identifiers by their scheme. This was the functionality I was expecting
from the method when I tried to use it.

The #wikidata method can still be used to return a 'wikidata' identifier, so
no functionality is lost. You can also use `person.identifier('wikidata')` to
access it.