Support access to Popolo Area objects
In PublicWhip we’re connecting memberships to their Areas. We pull this
information out of the EveryPolitician Popolo data using custom code.
https://github.com/OPORA/publicwhip/blob/develop/lib/data_loader/ukraine/people.rb#L45

We should be using this great gem for access all this data.

For consistency and simplicity, this gem should help us access Popolo Area data
like it does for Persons and Organizations :)