Remove leading components from IDs
EveryPolitician sometimes prefixes IDs with the 'type' of entity it's
referring to: "term/8", "party/abc" etc.

It's inconsistent with this (largely it depends on where the ID
originated), and it's largely a pain to work with. To make everything
simpler, we will strip those prefixes off when we see them.