convert index result to array before intersecting in where
If the index has no matching results then it is nil and not [].
Intersect throws an exception if you try to run it on something that
isn't an array so convert it to an array first.

Note that this will throw an exception if we get something that doesn't
have a to_a method but that should not happen and if it does then
something has gone wrong.

Fixes #69