check all attributes to determine membership equality
Previously membership equality testing did not work as it was using .id
to compare which was always empty and so everything matched. Now we make
sure that all attributes are equal.

This also fixes searching on memberships using the new index method as
that uses the intersect method which compares objects using .eql?

Fixes #70