use a factory to generated event classes
this adds an EntityFactory which EventFactory inherits from to configure
the subclasses and default class. This is then used as the entity_class
for Events.

This means we are creating the objects in an intermediate class which is
less confusing and fragile than the Event class returning objects which
are not Events. It also allows us to declare the possible Event classes
in a fairly straight forward way, and doesn't rely on meta-programming
to work it out. Finally, it should be adaptable to other places we want
to use this.