return a Collection from where method
This changes `Collection.where` to return another Collection based
object rather than an array. This is to allow further calls to `.where`
and `.find_by` to be done on the results. This is less important for
primary calls but more useful when the result is returned from methods
like `.terms`.

Fixes #97