Handle events with no start date
This means that calling `Everypolitician::Popolo::JSON#events` won't
raise an exception when an event is missing a `start_date` attribute.

Previously an exception was being raised when some events had a start
date and some didn't because `#sort_by` was trying to compare a string
with `nil`, which was causing the following error:

    ArgumentError: comparison of NilClass with String failed

By casting the `start_date` to a string in the `#sort_by` block we can
avoid this problem.